requirements.txt
django==4.2.4
djangorestframework==3.14.0
psycopg2-binary==2.9.7
markdown==3.4.4
django-filter==23.2
serpy==0.3.1
requests==2.31.0
django-cors-headers==4.2.0
django-health-check==3.17.0
drf-spectacular==0.26.4
reportlab==4.0.4
boto3==1.28.24